Move hot reloading into a separate module
